refactor: extract feature command registry
