+++ internal/il2cpp-dll-injection.vcxproj
@@ -87,6 +87,7 @@
     <ClCompile Include="src\features\movement\noclip\Noclip.cpp" />
     <ClCompile Include="src\features\movement\noclip\NoclipHook.cpp" />
     <ClCompile Include="src\features\control\FeatureState.cpp" />
+    <ClCompile Include="src\features\control\FeatureCommandRegistry.cpp" />
     <ClCompile Include="src\features\visuals\FloatingTextService.cpp" />
     <ClCompile Include="src\features\runtime\FeatureRuntime.cpp" />
     <ClCompile Include="vendor\minhook\hook.c">
@@ -222,6 +223,7 @@
     <ClInclude Include="src\features\movement\noclip\Noclip.h" />
     <ClInclude Include="src\features\movement\noclip\NoclipHook.h" />
     <ClInclude Include="src\features\control\FeatureState.h" />
+    <ClInclude Include="src\features\control\FeatureCommandRegistry.h" />
     <ClInclude Include="src\features\visuals\FloatingTextService.h" />
     <ClInclude Include="src\features\runtime\FeatureRuntime.h" />
     <ClInclude Include="src\game\symbols\BeebyteName.h" />
