+++ internal/src/core/ipc/IpcBridge.cpp
@@ -8,6 +8,7 @@
 #include "FeatureState.h"
 #include "FloatingTextService.h"
 #include "FeatureRuntime.h"
+#include "FeatureCommandRegistry.h"
 
 // Debug logging
 
@@ -52,38 +53,12 @@ bool        IpcBridge_IsAuthenticated() { return Handshake::IsHealthy(&s_auth); 
 static std::atomic<bool> s_overlayEnabled{false};
 
 bool IpcBridge_IsOverlayEnabled() { return s_overlayEnabled.load(std::memory_order_relaxed); }
-
-// Render-thread feature applicators — moved to FeatureRuntime.cpp
-
-namespace {
-
-    int ResolveHotkeyVk(const char* raw)
-    {
-        if (!raw) return 0;
-        std::string key;
-        for (const char* p = raw; *p; ++p) {
-            const unsigned char ch = static_cast<unsigned char>(*p);
-            if (!std::isspace(ch)) key.push_back(static_cast<char>(std::toupper(ch)));
-        }
-        if (key.empty() || key == "NONE" || key == "OFF") return 0;
-        if (key.rfind("VK_", 0) == 0) key.erase(0, 3);
-        if (key.size() == 1) {
-            const char ch = key[0];
-            if ((ch >= 'A' && ch <= 'Z') || (ch >= '0' && ch <= '9')) return static_cast<int>(ch);
-        }
-        if (key.size() >= 2 && key[0] == 'F') {
-            const int fn = atoi(key.c_str() + 1);
-            if (fn >= 1 && fn <= 24) return VK_F1 + (fn - 1);
-        }
-        if (key.rfind("NUMPAD", 0) == 0) {
-            const int n = atoi(key.c_str() + 6);
-            if (n >= 0 && n <= 9) return VK_NUMPAD0 + n;
-        }
-        static const std::pair<const char*, int> named[] = {{"SPACE", VK_SPACE}, {"TAB", VK_TAB}, {"ESC", VK_ESCAPE}, {"ESCAPE", VK_ESCAPE}, {"SHIFT", VK_SHIFT}, {"CTRL", VK_CONTROL}, {"CONTROL", VK_CONTROL}, {"ALT", VK_MENU}, {"MENU", VK_MENU}, {"INSERT", VK_INSERT}, {"INS", VK_INSERT}, {"DELETE", VK_DELETE}, {"DEL", VK_DELETE}, {"HOME", VK_HOME}, {"END", VK_END}, {"PAGEUP", VK_PRIOR}, {"PGUP", VK_PRIOR}, {"PAGEDOWN", VK_NEXT}, {"PGDN", VK_NEXT}, {"UP", VK_UP}, {"DOWN", VK_DOWN}, {"LEFT", VK_LEFT}, {"RIGHT", VK_RIGHT}};
-        for (const auto& kv : named) if (key == kv.first) return kv.second;
-        return 0;
-    }
-} // namespace
+void IpcBridge_SetOverlayEnabled(bool on)
+{
+    s_overlayEnabled.store(on, std::memory_order_relaxed);
+    if (!on) settings.bShowMenu = false;
+    DbgLog("overlayEnabled = %s", on ? "true" : "false");
+}
 
 // Public feature accessors — shim to FeatureState
 
@@ -237,33 +212,6 @@ static bool DispatchAuthMessage(char* json, HANDLE hPipe, char* msgBuf, int msgB
 
 // Feature command parsing and dispatch
 
-struct FeatureCommand {
-    char key[64] = {};
-    char valueType[8] = {};
-    char value[128] = {};
-
-    bool Is(const char* name) const { return strcmp(key, name) == 0; }
-    bool Bool() const { return strcmp(value, "true") == 0 || strcmp(value, "1") == 0; }
-    int Int() const { return atoi(value); }
-    float Float() const { return static_cast<float>(atof(value)); }
-};
-
-struct FeatureHandler { const char* key; bool (*apply)(const FeatureCommand&); };
-
-static bool ApplyFeatureTable(const FeatureCommand& f, const FeatureHandler* handlers, size_t count)
-{
-    for (size_t i = 0; i < count; ++i)
-        if (f.Is(handlers[i].key)) return handlers[i].apply(f);
-    return false;
-}
-
-#define FH(key, body) { key, [](const FeatureCommand& f)->bool { body; return true; } }
-#define FH_BOOL(key, fn) FH(key, fn(f.Bool()))
-#define FH_INT(key, fn) FH(key, fn(f.Int()))
-#define FH_INT_BOOL(key, fn) FH(key, fn(f.Int() != 0))
-#define FH_FLOAT(key, fn) FH(key, fn(f.Float()))
-#define FH_TEXT(key, fn) FH(key, fn(f.value))
-
 static bool ParseSetFeatureCommand(char* json, const char* seqStr, const char* macHex, FeatureCommand* out)
 {
     if (!out) return false;
@@ -312,159 +260,11 @@ static bool DispatchTileCommand(const char* type, char* json, const char* seqStr
     return false;
 }
 
-static bool ApplyCoreFeature(const FeatureCommand& f)
-{
-    static const FeatureHandler h[] = {
-        FH("overlayEnabled", {
-            const bool on = f.Bool();
-            s_overlayEnabled.store(on, std::memory_order_relaxed);
-            if (!on) settings.bShowMenu = false;
-            DbgLog("overlayEnabled = %s", on ? "true" : "false");
-        }),
-        FH("internalUnloadDll", {
-            if (f.Bool()) {
-                DBG_FILE_LOG("[IpcBridge] internalUnloadDll requested.");
-                IpcBridge_RequestShutdown();
-                if (hUnloadEvent) SetEvent(hUnloadEvent);
-            }
-        }),
-        FH_BOOL("autoAimEnabled", IpcBridge_SetAutoAimEnabled),
-        FH_INT("autoAimMode", IpcBridge_SetAutoAimMode),
-        FH_BOOL("autoAimPrioritizeBosses", AutoAim::SetPrioritizeBosses),
-        FH_BOOL("autoAimIgnoreWalls", AutoAim::SetIgnoreWalls),
-        FH("projectileNoclipEnabled", {
-            const bool on = f.Bool();
-            FeatureState::SetProjectileNoclipEnabled(on);
-            if (!on) ProjNoclip::SetEnabled(false);
-        }),
-        FH("clientDefense", FeatureState::SetClientDefense(static_cast<int32_t>(f.Int()))),
-        FH("clientClassType", FeatureState::SetClientClassType(static_cast<int32_t>(f.Int()))),
-        FH_INT("autoDodgeMode", IpcBridge_SetAutoDodgeMode),
-        FH_FLOAT("autoDodgeHorizonMs", IpcBridge_SetAutoDodgeHorizonMs),
-        FH_FLOAT("autoDodgeHitboxPadding", IpcBridge_SetAutoDodgeHitboxPadding),
-        FH_BOOL("autoDodgeWallAvoid", IpcBridge_SetAutoDodgeWallAvoid),
-        FH("gameHitboxMult", TestTAB::SetGameHitboxOverride(f.Float() < 1.0f - 1e-4f, f.Float())),
-        FH_FLOAT("speedHackMult", SpeedHack::SetMultiplier),
-        FH_BOOL("autoAbilityEnabled", IpcBridge_SetAutoAbilityEnabled),
-        FH_FLOAT("autoAbilityMpPct", IpcBridge_SetAutoAbilityMpPct),
-        FH_INT("autoAbilityWizardMode", IpcBridge_SetAutoAbilityWizardMode),
-        FH_INT("targetFrameRate", FpsSetter::SetTargetFps),
-        FH_TEXT("showPluginFloatingText", FloatingTextService::QueuePluginText)
-    };
-    return ApplyFeatureTable(f, h, sizeof(h) / sizeof(h[0]));
-}
-
-static bool ApplyXDodgeFeature(const FeatureCommand& f)
-{
-    static const FeatureHandler h[] = {
-        FH_FLOAT("xdodgeHitScale", XDodge::SetHitScale),
-        FH_INT("xdodgeRebuildN", XDodge::SetRebuildN),
-        FH_FLOAT("xdodgePlanStepMs", XDodge::SetPlanStepMs),
-        FH_FLOAT("xdodgeSearchRadius", XDodge::SetSearchRadius),
-        FH_INT_BOOL("xdodgeAstar", XDodge::SetAstarEnabled),
-        FH_INT_BOOL("xdodgeWeighting", XDodge::SetWeightingEnabled),
-        FH_INT_BOOL("xdodgeSmartGoal", XDodge::SetSmartGoalEnabled),
-        FH_INT_BOOL("xdodgePerpBias", XDodge::SetPerpEnabled),
-        FH_INT_BOOL("xdodgeSpeedMatch", XDodge::SetSpeedMatchEnabled),
-        FH_INT_BOOL("xdodgeLockFollow", DangerPlanner::SetLockFollowEnabled),
-        FH_INT("xdodgeAutoLock", DangerPlanner::SetAutoLockMode),
-        FH_INT_BOOL("xdodgeWalkCache", XDodge::SetWalkCacheEnabled),
-        FH_INT_BOOL("xdodgeWallAvoid", XDodge::SetWallAvoidEnabled),
-        FH_INT_BOOL("xdodgeArbiter", XDodge::SetArbiterEnabled),
-        FH_INT_BOOL("xdodgeBfsBias", XDodge::SetBfsBiasEnabled),
-        FH_INT_BOOL("xdodgeCcd", XDodge::SetCcdEnabled),
-        FH_FLOAT("xdodgeCcdPad", XDodge::SetCcdPad),
-        FH_INT_BOOL("xdodgeCatalog", XDodge::SetCatalogEnabled),
-        FH("xdodgeNotifyHit", XDodge::OnPlayerHit()),
-        FH_INT_BOOL("xdodgeDrawPath", XDodge::SetDrawPathEnabled),
-        FH_INT_BOOL("xdodgeDrawProjPred", XDodge::SetDrawProjPredEnabled),
-        FH_FLOAT("xdodgeDebugPredLongMs", XDodge::SetDebugPredLongMs),
-        FH_INT_BOOL("xdodgeAvoidEnemies", XDodge::SetAvoidEnemiesEnabled),
-        FH_INT_BOOL("xdodgeGhostHit", GhostHit::SetEnabled),
-        FH_INT_BOOL("xdodgeLosGoal", XDodge::SetLosGoalEnabled),
-        FH_INT_BOOL("xdodgeWasdYield", XDodge::SetWasdYieldEnabled),
-        FH_INT_BOOL("xdodgeLateralPref", XDodge::SetLateralPrefEnabled),
-        FH_INT_BOOL("xdodgeGoalSticky", XDodge::SetGoalStickyEnabled),
-        FH_FLOAT("xdodgeDangerPenalty", XDodge::SetDangerWeight)
-    };
-    if (ApplyFeatureTable(f, h, sizeof(h) / sizeof(h[0]))) return true;
-    if (f.Is("xdodgeStayPenalty") || f.Is("xdodgeFutureSample") || f.Is("xdodgeFutureHorizon") || f.Is("xdodgeFutureStride")) return true;
-    return false;
-}
-
-static bool ApplyRolloutFeature(const FeatureCommand& f)
-{
-    static const FeatureHandler h[] = {
-        FH_FLOAT("rolloutHorizonTicks", RolloutDodge::SetHorizonTicks),
-        FH_FLOAT("rolloutSampleStepMs", RolloutDodge::SetSampleStepMs),
-        FH_INT("rolloutHeadings", RolloutDodge::SetHeadingCount),
-        FH_FLOAT("rolloutHitScale", RolloutDodge::SetHitScale),
-        FH_FLOAT("rolloutIntentWeight", RolloutDodge::SetIntentWeight),
-        FH_INT("rolloutRebuildN", RolloutDodge::SetRebuildN),
-        FH_INT_BOOL("rolloutForceBrute", RolloutDodge::SetForceBruteForce),
-        FH_INT_BOOL("rolloutAvoidEnemies", RolloutDodge::SetAvoidEnemiesEnabled),
-        FH_INT_BOOL("rolloutWasdYield", RolloutDodge::SetWasdYieldEnabled),
-        FH_INT_BOOL("rolloutCommitDwell", RolloutDodge::SetCommitDwellEnabled),
-        FH_INT_BOOL("rolloutDrawPath", RolloutDodge::SetDrawPathEnabled)
-    };
-    return ApplyFeatureTable(f, h, sizeof(h) / sizeof(h[0]));
-}
-
-static bool ApplyInputCameraSkinFeature(const FeatureCommand& f)
-{
-    static const FeatureHandler h[] = {
-        FH("playerNoclipActive",      FeatureState::SetPlayerNoclipActive(f.Bool())),
-        FH("playerNoclipEnabled",     FeatureState::SetPlayerNoclipEnabled(f.Bool())),
-        FH("playerNoclipHotkey",      FeatureState::SetPlayerNoclipHotkeyVk(ResolveHotkeyVk(f.value))),
-        FH("socketHotkeyActive",      FeatureState::SetSocketHotkeyActive(f.Bool())),
-        FH("socketHotkey",            FeatureState::SetSocketHotkeyVk(ResolveHotkeyVk(f.value))),
-        FH("walkTargetX",             FeatureState::SetWalkTarget(f.Float(), FeatureState::GetWalkTargetY(), FeatureState::GetWalkTargetActive())),
-        FH("walkTargetY",             FeatureState::SetWalkTarget(FeatureState::GetWalkTargetX(), f.Float(), FeatureState::GetWalkTargetActive())),
-        FH("walkTargetActive",        FeatureState::SetWalkTarget(FeatureState::GetWalkTargetX(), FeatureState::GetWalkTargetY(), f.Bool())),
-        FH("cameraZoomActive",        FeatureState::SetCameraZoom(f.Bool(), FeatureState::GetCameraZoomValue())),
-        FH("cameraZoomValue",         FeatureState::SetCameraZoom(FeatureState::GetCameraZoomActive(), f.Float())),
-        FH("cameraAngleActive",       FeatureState::SetCameraAngle(f.Bool(), FeatureState::GetCameraAngleValue())),
-        FH("cameraAngleValue",        FeatureState::SetCameraAngle(FeatureState::GetCameraAngleActive(), f.Int())),
-        FH("cameraCenteringActive",   FeatureState::SetCameraCentering(f.Bool(), FeatureState::GetCameraCentered())),
-        FH("cameraCentered",          FeatureState::SetCameraCentering(FeatureState::GetCameraCenteringActive(), f.Bool())),
-        FH("skinOverrideEnabled",     FeatureState::SetSkinOverride(f.Bool(), FeatureState::GetSkinOverrideId())),
-        FH("skinOverrideId",          FeatureState::SetSkinOverride(FeatureState::GetSkinOverrideEnabled(), f.Int()))
-    };
-    return ApplyFeatureTable(f, h, sizeof(h) / sizeof(h[0]));
-}
-
-static bool ApplyDangerPlannerFeature(const FeatureCommand& f)
-{
-    static const FeatureHandler h[] = {
-        FH_FLOAT("dodgeWasdLookahead", DangerPlanner::SetWasdLookahead),
-        FH_FLOAT("dodgeTightLeash", DangerPlanner::SetTightLeashRadius),
-        FH_FLOAT("dodgeIdleMinGain", DangerPlanner::SetIdleMinGain),
-        FH_FLOAT("dodgeStickiness", DangerPlanner::SetStickiness),
-        FH_BOOL("dodgeReplanOnSpawn", DangerPlanner::SetReplanOnHazardSpawn),
-        FH_BOOL("dodgeStrategicBias", DangerPlanner::SetStrategicBiasEnabled),
-        FH_BOOL("dodgeStrategicNearWaypoint", DangerPlanner::SetStrategicUseNearWaypoint),
-        FH_FLOAT("dodgeHitAversion", DangerPlanner::SetHitAversion),
-        FH_FLOAT("dodgeHitScale", DangerPlanner::SetDodgeHitScale)
-    };
-    return ApplyFeatureTable(f, h, sizeof(h) / sizeof(h[0]));
-}
-
-#undef FH_TEXT
-#undef FH_FLOAT
-#undef FH_INT_BOOL
-#undef FH_INT
-#undef FH_BOOL
-#undef FH
-
 static void DispatchSetFeature(char* json, const char* seqStr, const char* macHex)
 {
     FeatureCommand feature{};
     if (!ParseSetFeatureCommand(json, seqStr, macHex, &feature)) return;
-    if (ApplyCoreFeature(feature)) return;
-    if (ApplyXDodgeFeature(feature)) return;
-    if (ApplyRolloutFeature(feature)) return;
-    if (ApplyInputCameraSkinFeature(feature)) return;
-    ApplyDangerPlannerFeature(feature);
+    FeatureCommandRegistry::Apply(feature);
 }
 
 static void DispatchCommand(char* json)
